# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI智能助手系统",
    "summary": "AI智能助手系统API",
    "description": "\n## AI智能助手系统后端API\n\n基于FastAPI和LangChain 1.0框架构建的企业级AI智能助手系统，提供以下核心功能：\n\n### 核心功能\n\n* **智能对话** - 支持多轮对话、流式响应、对话历史管理\n* **RAG知识库** - 文档上传、向量化、智能检索问答\n* **Agent智能代理** - 工具调用、任务执行、推理过程可视化\n* **用户管理** - 注册登录、JWT认证、配额管理\n* **系统监控** - 使用统计、健康检查、Prometheus指标\n\n### 技术栈\n\n* **Web框架**: FastAPI 0.104+\n* **AI框架**: LangChain 1.0\n* **数据库**: MySQL 8.0, Redis 7.0, Chroma\n* **LLM**: 通义千问 (Tongyi Qianwen)\n\n### 认证方式\n\n大部分API端点需要JWT认证。在请求头中添加：\n```\nAuthorization: Bearer <access_token>\n```\n\n### 速率限制\n\n* 登录接口: 5次/分钟\n* 普通API: 100次/分钟\n* LLM调用: 20次/分钟\n\n### 错误响应格式\n\n所有错误响应遵循统一格式：\n```json\n{\n    \"error_code\": \"1001\",\n    \"message\": \"错误描述\",\n    \"request_id\": \"uuid\"\n}\n```\n    ",
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "API Support",
      "url": "https://example.com/support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/rag/query": {
      "post": {
        "tags": [
          "RAG问答"
        ],
        "summary": "RAG问答",
        "description": "基于知识库进行智能问答，支持单个或多个知识库联合检索。",
        "operationId": "rag_query_api_v1_rag_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAGQueryResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/rag/query/stream": {
      "post": {
        "tags": [
          "RAG问答"
        ],
        "summary": "RAG流式问答",
        "description": "基于知识库进行流式智能问答，实时返回生成的答案。",
        "operationId": "rag_query_stream_api_v1_rag_query_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "参数验证失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "401": {
            "description": "未授权 - JWT令牌无效或已过期",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "1002",
                  "message": "令牌已过期",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "403": {
            "description": "禁止访问 - 配额不足或权限不足",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "4001",
                  "message": "配额已用尽",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "2001",
                  "message": "对话不存在",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁 - 超出速率限制",
            "content": {
              "application/json": {
                "example": {
                  "error": "Rate limit exceeded",
                  "message": "请求过于频繁，请稍后再试"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "3001",
                  "message": "数据库连接失败",
                  "request_id": "550e8400-e29b-41d4-a716-446655440000"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentChunkResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content",
            "description": "片段内容"
          },
          "document_name": {
            "type": "string",
            "title": "Document Name",
            "description": "来源文档名称"
          },
          "similarity_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Similarity Score",
            "description": "相似度评分"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id",
            "description": "文档ID"
          },
          "chunk_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Index",
            "description": "分块索引"
          }
        },
        "type": "object",
        "required": [
          "content",
          "document_name",
          "similarity_score"
        ],
        "title": "DocumentChunkResponse",
        "description": "文档片段响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RAGQueryRequest": {
        "properties": {
          "knowledge_base_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Knowledge Base Ids",
            "description": "知识库ID列表"
          },
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "description": "查询问题"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "description": "检索文档数量",
            "default": 5
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id",
            "description": "对话ID（用于维护上下文）"
          }
        },
        "type": "object",
        "required": [
          "knowledge_base_ids",
          "question"
        ],
        "title": "RAGQueryRequest",
        "description": "RAG查询请求"
      },
      "RAGQueryResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer",
            "description": "生成的答案"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/DocumentChunkResponse"
            },
            "type": "array",
            "title": "Sources",
            "description": "参考文档片段"
          },
          "tokens_used": {
            "type": "integer",
            "title": "Tokens Used",
            "description": "消耗的token数量"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "sources",
          "tokens_used"
        ],
        "title": "RAGQueryResponse",
        "description": "RAG查询响应"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}